refac: move AdminService to the admin  package
